+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ## Bug fixes
 - Fixed a bug that caused the funds to not be asked when needed
 - Fixed the visualization of poll results
-- Fixed a bug that causes a wrong visualization of a post details
+- Fixed a bug that caused a wrong visualization of a post details
 
 # Version 0.4.1
 ## Changes